FileFinder_19/pyproject.toml
[tool.poetry]
name = "filefinder"
version = "0.1.0"
description = ""
authors = ["Arun <arunkg99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
altgraph = "0.17.4"
build = "1.0.3"
cachecontrol = "0.13.1"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
distlib = "0.3.7"
dulwich = "0.21.6"
et-xmlfile = "1.1.0"
fastjsonschema = "2.19.0"
filelock = "3.13.1"
idna = "3.4"
importlib-metadata = "6.8.0"
installer = "0.7.0"
jaraco-classes = "3.3.0"
keyring = "24.3.0"
markdown-it-py = "3.0.0"
mdurl = "0.1.2"
more-itertools = "10.1.0"
msgpack = "1.0.7"
mysql-connector-python = "8.2.0"
numpy = "1.26.1"
openpyxl = "3.1.2"
packaging = "23.2"
pandas = "2.1.2"
pefile = "2023.2.7"
pexpect = "4.8.0"
pkginfo = "1.9.6"
platformdirs = "3.11.0"
poetry = "1.7.1"
poetry-core = "1.8.1"
poetry-plugin-export = "1.6.0"
prompt-toolkit = "3.0.36"
protobuf = "4.21.12"
psutil = "5.9.6"
ptyprocess = "0.7.0"
pygments = "2.17.2"
pyinstaller = "6.1.0"
pyinstaller-hooks-contrib = "2023.10"
pyproject-hooks = "1.0.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pywin32-ctypes = "0.2.2"
questionary = "2.0.1"
rapidfuzz = "3.5.2"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
rich = "13.7.0"
shellingham = "1.5.4"
six = "1.16.0"
tomlkit = "0.12.3"
trove-classifiers = "2023.11.22"
tzdata = "2023.3"
urllib3 = "2.1.0"
virtualenv = "20.24.7"
wcwidth = "0.2.12"
xlrd = "2.0.1"
zipp = "3.17.0"
keyboard = "0.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
